pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenttrader"
version = "0.1.0"
description = "Prediction market strategy research, backtesting, and paper trading toolkit"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "agenttrader" }]
dependencies = [
  "dome-api-sdk>=1.0.0",
  "click>=8.1",
  "mcp>=1.0.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pyyaml>=6.0",
  "watchdog>=4.0",
  "msgpack>=1.0",
  "tenacity>=8.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "websockets>=12.0",
  "rich>=13.0",
  "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.scripts]
agenttrader = "agenttrader.cli.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["agenttrader*"]

[tool.setuptools.package-data]
agenttrader = ["dashboard/static/*"]
